refactor AddTodo & Footer & FilterLink
