+++ redux-demo/src/index.js
@@ -3,27 +3,9 @@ import ReactDOM from 'react-dom';
 
 import store from './todosReducer';
 import TodoList from './TodoList';
+import AddTodo from './AddTodo';
+import Footer from './Footer';
 
-const FilterLink = ({
-  filter,
-  currentFilter,
-  children
-}) => {
-  if (filter === currentFilter) {
-    return <span>{children}</span>
-  }
-  return <a href='#'
-    onClick={e => {
-      e.preventDefault();
-      store.dispatch({
-        type: 'SET_VISIBILITY_FILTER',
-        filter
-      });
-    }}
-  >
-    {children}
-  </a>
-}
 
 const getVisibleTodos = (
   todos,
@@ -44,67 +26,47 @@ const getVisibleTodos = (
 }
 
 let nextTodoId = 0;
-class TodoApp extends React.Component {
-  render() {
-    const {
+const TodoApp = ({
       todos,
-      visibilityFilter
-    } = this.props;
-    const visibleTodos = getVisibleTodos(
-      todos,
-      visibilityFilter
-    );
-    return (
-      <div>
-        <input ref={node => {
-          this.input = node;
-        }} />
-        <button onClick={() => {
+  visibilityFilter,
+}) => {
+  return (
+    <div>
+      <AddTodo
+        onAddClick={text =>
           store.dispatch({
             type: 'ADD_TODO',
-            text: this.input.value,
+            text,
             id: nextTodoId++
-          });
-          this.input.value = '';
-        }}>
-          Add Todo
-        </button>
-        <TodoList
-          todos={visibleTodos}
-          onTodoClick={id => {
+          })
+        }
+      />
+      <TodoList
+        todos={
+          getVisibleTodos(
+            todos,
+            visibilityFilter
+          )}
+        onTodoClick={
+          id => {
             store.dispatch({
               type: 'TOGGLE_TODO',
               id,
             });
-          }}
-        />
-        <p>
-          Show:
-          {' '}
-          <FilterLink
-            filter='SHOW_ALL'
-            currentFilter={visibilityFilter}
-          >
-            All
-          </FilterLink>
-          {' '}
-          <FilterLink
-            filter='SHOW_ACTIVE'
-            currentFilter={visibilityFilter}
-          >
-            Active
-          </FilterLink>
-          {' '}
-          <FilterLink
-            filter='SHOW_COMPLETED'
-            currentFilter={visibilityFilter}
-          >
-            Completed
-          </FilterLink>
-        </p>
-      </div>
-    )
-  };
+          }
+        }
+      />
+      <Footer
+        visibilityFilter={visibilityFilter}
+        onFilterClick={filter =>
+          store.dispatch({
+            type: 'SET_VISIBILITY_FILTER',
+            filter
+          })
+        }
+      />
+    </div>
+  )
 }
 
 const render = () => {
